requirements.txt
adrf==0.1.4
asgiref==3.7.2
async-property==0.2.2
channels==4.0.0
Django==4.2.10
djangorestframework==3.14.0
drf-yasg==1.21.7
ffmpeg-python==0.2.0
future==0.18.3
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
